refactor(manufacture): extract FefoConsumptionAllocator
